[WS-252] Fixes argument mismatches for minecraft 1.8 (portals) * Updates controll buttons after an animation has been added to a portal

+++ src/de/codingair/warpsystem/spigot/features/portals/guis/subgui/HologramEditor.java
@@ -19,7 +19,7 @@ import de.codingair.warpsystem.spigot.base.WarpSystem;
 import de.codingair.warpsystem.spigot.base.language.Lang;
 import de.codingair.warpsystem.spigot.features.playerwarps.managers.PlayerWarpManager;
 import de.codingair.warpsystem.spigot.features.portals.guis.PortalEditor;
-import de.codingair.warpsystem.spigot.features.portals.guis.animationseditor.AnimationHotBarEditor;
+import de.codingair.warpsystem.spigot.features.portals.guis.subgui.animationseditor.AnimationHotBarEditor;
 import de.codingair.warpsystem.spigot.features.portals.utils.Hologram;
 import net.md_5.bungee.api.chat.BaseComponent;
 import net.md_5.bungee.api.chat.ClickEvent;
@@ -145,7 +145,7 @@ public class HologramEditor extends HotbarGUI {
         } else {
             //loc, mat, byte
             IReflection.MethodAccessor sendBlockChange = IReflection.getMethod(Player.class, "sendBlockChange", null, new Class[] {org.bukkit.Location.class, Material.class, byte.class});
-            sendBlockChange.invoke(player, loc, XMaterial.GLASS.parseMaterial(), 1);
+            sendBlockChange.invoke(player, loc, XMaterial.GLASS.parseMaterial(), (byte) 1);
         }
     }
